// backend/src/modules/user/schemas/create-user.schema.ts
import { UserRole } from "@prisma/client"
import { z } from "zod"

export const createUserSchema = z.object({
  name: z.string().min(3, "Nome deve ter pelo menos 3 caracteres"),
  email: z.string().email("Email inválido"),
  password: z.string().min(6, "Senha deve ter pelo menos 6 caracteres"),
  role: z.nativeEnum(UserRole) //impede do usuario criar um perfil invalido como GERENTE
})